EmailService added and notification sent to email
