+++ pom.xml
@@ -54,16 +54,11 @@
             <version>0.11.5</version>
         </dependency>
 
-        <dependency>
-            <groupId>javax.mail</groupId>
-            <artifactId>mail</artifactId>
-            <version>1.4</version>
-        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-mail</artifactId>
-            <version>3.2.3</version>
+            <version>3.2.2</version>
         </dependency>
 
 
@@ -81,7 +76,7 @@
         <dependency>
             <groupId>org.modelmapper</groupId>
             <artifactId>modelmapper</artifactId>
-            <version>2.4.4</version> <!-- Make sure to use the appropriate version -->
+            <version>2.4.4</version>
         </dependency>
 
     </dependencies>
