+++ src/main/java/az/code/tensapi/TensApiApplication.java
@@ -9,16 +9,10 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableScheduling
-public class TensApiApplication implements CommandLineRunner {
-    @Autowired
-    TaskServiceImpl taskService;
+public class TensApiApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(TensApiApplication.class, args);
     }
 
-    @Override
-    public void run(String... args) throws Exception {
-        taskService.sendTaskNotification();
-    }
 }
