+++ src/main/java/az/code/tensapi/service/impl/EmailServiceImpl.java
@@ -3,13 +3,14 @@ package az.code.tensapi.service.impl;
 
 import az.code.tensapi.service.EmailService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
-
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class EmailServiceImpl implements EmailService {
     private final JavaMailSender javaMailSender;
 
@@ -26,7 +27,8 @@ public class EmailServiceImpl implements EmailService {
             message.setText(notification);
             javaMailSender.send(message);
         } catch (Exception e) {
-            throw new RuntimeException(e.getMessage());
+            log.error("An error occurred while sending the email: {}", e.getMessage());
+            throw new RuntimeException("An error occurred while sending the email. Please try again later.");
         }
     }
 }
