+++ src/main/resources/application.yml
@@ -21,6 +21,7 @@ spring:
           auth: true
           starttls:
             enable: true
+
 jwt-secret: 8f09669ab33ae0c90f375f07bca6b22f5ac2a496e505c5ffa3cf15460b69a882
 
 openai:
@@ -28,3 +29,4 @@ openai:
   api:
     key: ${api_key}
     url: "https://api.openai.com/v1/chat/completions"
+
